Move numeroSuggestion in editableMarker
