+++ components/map/editable-marker.js
@@ -1,19 +1,56 @@
-import React, {useCallback, useContext} from 'react'
+import React, {useCallback, useContext, useState} from 'react'
 import PropTypes from 'prop-types'
 import {Marker} from 'react-map-gl'
 import {Pane, MapMarkerIcon, Text} from 'evergreen-ui'
+import nearestPointOnLine from '@turf/nearest-point-on-line'
+import length from '@turf/length'
+import lineSlice from '@turf/line-slice'
 
 import MarkersContext from '../../contexts/markers'
 
-function EditableMarker({size, style}) {
-  const {markers, updateMarker} = useContext(MarkersContext)
+function EditableMarker({size, style, voie}) {
+  const {markers, updateMarker, setOverrideText} = useContext(MarkersContext)
+  const [suggestedNumero, setSuggestedNumero] = useState(null)
 
   const onDragEnd = useCallback((event, idx) => {
     const [longitude, latitude] = event.lngLat
     const {_id, type} = markers[idx]
 
+    setOverrideText(suggestedNumero)
     updateMarker(_id, {longitude, latitude, type})
-  }, [markers, updateMarker])
+  }, [markers, updateMarker, setOverrideText, suggestedNumero])
+
+  const onDrag = useCallback(event => {
+    const [longitude, latitude] = event.lngLat
+
+    const numeroSuggestion = () => {
+      if (voie.trace) {
+        const {trace} = voie
+        const point = {
+          type: 'Feature',
+          properties: {},
+          geometry: {
+            type: 'Point',
+            coordinates: [longitude, latitude]
+          }
+        }
+        const from = {
+          type: 'Feature',
+          properties: {},
+          geometry: {
+            type: 'Point',
+            coordinates: trace.coordinates[0]
+          }
+        }
+
+        const to = nearestPointOnLine({type: 'Feature', geometry: trace}, point, {units: 'kilometers'})
+        const slicedLine = length(lineSlice(from, to, trace)) * 1000
+        return slicedLine.toFixed(0)
+      }
+    }
+
+    setSuggestedNumero(numeroSuggestion)
+  }, [voie])
 
   if (markers.length === 0) {
     return null
@@ -25,6 +62,7 @@ function EditableMarker({size, style}) {
         key={marker._id}
         {...marker}
         draggable
+        onDrag={e => onDrag(e)}
         onDragEnd={e => onDragEnd(e, idx)}
       >
         <Pane>
@@ -40,7 +78,7 @@ function EditableMarker({size, style}) {
             fontSize={10}
             whiteSpace='nowrap'
           >
-            {marker.type}
+            {suggestedNumero && `${suggestedNumero} - `}{marker.type}
           </Text>
 
           <MapMarkerIcon
